cxgbe/iw_cxgbe: Remove another bad cast.  This should have been included in r316571.
MFC after:	3 days
Sponsored by:	Chelsio Communications

+++ sys/dev/cxgbe/iw_cxgbe/mem.c
@@ -443,7 +443,7 @@ int c4iw_reregister_phys_mem(struct ib_mr *mr, int mr_rereg_mask,
 		mhp->attr.zbva = 0;
 		mhp->attr.va_fbo = *iova_start;
 		mhp->attr.page_size = shift - 12;
-		mhp->attr.len = (u32) total_size;
+		mhp->attr.len = total_size;
 		mhp->attr.pbl_size = npages;
 	}
